src/lesson7/junit/Cat.java
package lesson7.junit;

/**
 * Created by alpo123 on 08.05.16.
 */
public class Cat {

    private String name;
    private double weight;
    private boolean isHungry = true;

    public Cat() {
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public double getWeight() {
        return weight;
    }

    public void setWeight(double weight) {
        this.weight = weight;
    }

    public boolean isHungry() {
        return isHungry;
    }

    public void setHungry(boolean hungry) {
        isHungry = hungry;
    }
}
